requirements.txt.txt
flask
flask_sqlalchemy
flask_bcrypt
flask_cors
pyjwt